# openapi.json
{
  "components" : {
    "schemas" : {
      "AllegroContext" : {
        "properties" : {
          "auctionId" : {
            "type" : "string"
          },
          "auctionName" : {
            "type" : "string"
          },
          "buyerLogin" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AskRequest" : {
        "properties" : {
          "conversationId" : {
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FaqEntryDto" : {
        "properties" : {
          "answer" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "displayOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "type" : "string"
          },
          "published" : {
            "type" : "boolean"
          },
          "question" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FaqEntryRequest" : {
        "properties" : {
          "answer" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "displayOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "published" : {
            "type" : "boolean"
          },
          "question" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageContent" : {
        "properties" : {
          "content" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ParticipationData" : {
        "properties" : {
          "arrivalDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "info" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "privacyPolicyAccepted" : {
            "type" : "boolean"
          },
          "registeredBy" : {
            "type" : "string"
          },
          "registrationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "size" : {
            "type" : "string"
          },
          "surveySendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "ticketSendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "voucher" : {
            "$ref" : "#/components/schemas/Voucher"
          }
        },
        "type" : "object"
      },
...
      "Presentation" : {
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "durationInMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "expectedPrice" : {
            "format" : "double",
            "type" : "number"
          },
          "id" : {
            "type" : "string"
          },
          "isFree" : {
            "type" : "boolean"
          },
          "language" : {
            "minLength" : 1,
            "type" : "string"
          },
          "level" : {
            "minLength" : 1,
            "type" : "string"
          },
          "maxGroupSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "new" : {
            "type" : "boolean"
          },
          "preSelectionComment" : {
            "type" : "string"
          },
          "preSelectionStatus" : {
            "enum" : [ "NONE", "PRE_APPROVED", "PRE_REJECTED", "IN_RESERVE" ],
            "type" : "string"
          },
          "ratingEnabled" : {
            "type" : "boolean"
          },
          "ratings" : {
            "items" : {
              "$ref" : "#/components/schemas/Rate"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "shortDescription" : {
            "minLength" : 1,
            "type" : "string"
          },
          "speaker" : {
            "$ref" : "#/components/schemas/Presentation",
            "writeOnly" : true
          },
          "speakers" : {
            "items" : {
              "$ref" : "#/components/schemas/User"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "status" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "title" : {
            "minLength" : 1,
            "type" : "string"
          },
          "workshop" : {
            "type" : "boolean"
          }
        },
        "required" : [ "description", "language", "level", "shortDescription", "speakers", "title" ],
        "type" : "object"
      },
...
      "PublicSpeaker" : {
        "properties" : {
          "bio" : {
            "type" : "string"
          },
          "github" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "photo" : {
            "type" : "string"
          },
          "presentations" : {
            "items" : {
              "$ref" : "#/components/schemas/Presentation"
            },
            "type" : "array"
          },
          "twitter" : {
            "type" : "string"
          },
          "www" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Rate" : {
        "properties" : {
          "comment" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "presentation" : {
            "$ref" : "#/components/schemas/Presentation"
          },
          "reviewerToken" : {
            "type" : "string"
          },
          "value" : {
            "enum" : [ "AWESOME", "GREAT", "IT_WAS_FINE", "BAD", "TERRIBLE" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReorderRequest" : {
        "properties" : {
          "ids" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "SseEmitter" : {
        "properties" : {
          "timeout" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Tag" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "User" : {
        "properties" : {
          "admin" : {
            "type" : "boolean"
          },
          "bio" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "github" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "origin" : {
            "type" : "string"
          },
          "participant" : {
            "type" : "boolean"
          },
          "participationData" : {
            "$ref" : "#/components/schemas/ParticipationData"
          },
          "photo" : {
            "type" : "string"
          },
          "privacyPolicyAccepted" : {
            "type" : "boolean"
          },
          "socialId" : {
            "type" : "string"
          },
          "speaker" : {
            "type" : "boolean"
          },
          "twitter" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "volunteer" : {
            "type" : "boolean"
          },
          "www" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Voucher" : {
        "properties" : {
          "allegro" : {
            "$ref" : "#/components/schemas/AllegroContext"
          },
          "comment" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "emailSent" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "originalBuyer" : {
            "type" : "string"
          },
          "ticketSendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "PARTICIPANT", "SPEAKER", "SPONSOR" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "Jelatyna API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/allegro/callback" : {
      "get" : {
        "operationId" : "callback_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "state",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "allegro-import-controller" ]
      }
    },
...
    "/chat/ask" : {
      "post" : {
        "operationId" : "ask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/event-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SseEmitter"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "chat-controller" ]
      }
    },
...
    "/faq-entries" : {
      "get" : {
        "operationId" : "getPublishedFaqEntries",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/FaqEntryDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      },
      "post" : {
        "operationId" : "createFaqEntry",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FaqEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FaqEntryDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      }
    },
...
    "/faq-entries/order" : {
      "put" : {
        "operationId" : "reorderFaqEntries",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReorderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      }
    },
    "/faq-entries/{id}" : {
      "delete" : {
        "operationId" : "deleteFaqEntry",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      },
      "get" : {
        "operationId" : "getFaqEntry",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FaqEntryDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      },
      "put" : {
        "operationId" : "updateFaqEntry",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FaqEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FaqEntryDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "faq-entry-controller" ]
      }
    },
...
    "/login/{provider}" : {
      "get" : {
        "operationId" : "loginInProvider",
        "parameters" : [ {
          "in" : "path",
          "name" : "provider",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "state",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "o-auth-2-login-controller" ]
      }
    },
    "/login/{provider}/callback" : {
      "get" : {
        "operationId" : "callback",
        "parameters" : [ {
          "in" : "path",
          "name" : "provider",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "state",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "o-auth-2-login-controller" ]
      }
    },
...
    "/pages/{id}" : {
      "delete" : {
        "operationId" : "deletePage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "page-controller" ]
      },
      "get" : {
        "operationId" : "getPage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "page-controller" ]
      },
      "post" : {
        "operationId" : "createPage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PageContent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "page-controller" ]
      },
      "put" : {
        "operationId" : "updatePage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PageContent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "page-controller" ]
      }
    },
...
    "/published/users/{id}/public" : {
      "get" : {
        "operationId" : "speaker",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicSpeaker"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "published-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}